DROID-2797 refactoring
