+++ app/src/main/java/com/anytypeio/anytype/di/main/UtilModule.kt
@@ -13,18 +13,19 @@ import com.anytypeio.anytype.core_utils.tools.DefaultUrlValidator
 import com.anytypeio.anytype.core_utils.tools.FeatureToggles
 import com.anytypeio.anytype.core_utils.tools.ThreadInfo
 import com.anytypeio.anytype.core_utils.tools.UrlValidator
+import com.anytypeio.anytype.device.providers.DateProviderImpl
 import com.anytypeio.anytype.domain.config.Gateway
 import com.anytypeio.anytype.domain.debugging.DebugConfig
 import com.anytypeio.anytype.domain.debugging.Logger
 import com.anytypeio.anytype.domain.misc.DateProvider
 import com.anytypeio.anytype.domain.misc.DateTypeNameProvider
+import com.anytypeio.anytype.domain.misc.LocaleProvider
 import com.anytypeio.anytype.domain.misc.UrlBuilder
 import com.anytypeio.anytype.middleware.interactor.MiddlewareProtobufLogger
 import com.anytypeio.anytype.middleware.interactor.ProtobufConverterProvider
 import com.anytypeio.anytype.other.BasicLogger
 import com.anytypeio.anytype.other.DefaultDateTypeNameProvider
 import com.anytypeio.anytype.other.DefaultDebugConfig
-import com.anytypeio.anytype.presentation.widgets.collection.DateProviderImpl
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
@@ -55,8 +56,12 @@ object UtilModule {
     @JvmStatic
     @Provides
     @Singleton
-    fun provideDateProvider(): DateProvider =
-        DateProviderImpl(defaultZoneId = ZoneId.systemDefault()
+    fun provideDateProvider(
+        localeProvider: LocaleProvider
+    ): DateProvider =
+        DateProviderImpl(
+            defaultZoneId = ZoneId.systemDefault(),
+            localeProvider = localeProvider
         )
 
     @Module
