+++ device/src/main/java/com/anytypeio/anytype/device/providers/DateProviderImpl.kt
@@ -1,4 +1,4 @@
-package com.anytypeio.anytype.persistence.providers
+package com.anytypeio.anytype.device.providers
 
 import android.text.format.DateUtils
 import com.anytypeio.anytype.core_models.TimeInMillis
@@ -139,7 +139,7 @@ class DateProviderImpl @Inject constructor(
             val formatter = SimpleDateFormat(pattern, locale)
             return formatter.format(Date(timestamp))
         } catch (e: Exception) {
-            Timber.e(e,"Error formatting timestamp to date string")
+            //Timber.e(e,"Error formatting timestamp to date string")
             return ""
         }
     }
@@ -160,7 +160,7 @@ class DateProviderImpl @Inject constructor(
             val timeString = timeFormatter.format(date)
             Pair(dateString, timeString)
         } catch (e: Exception) {
-            Timber.e(e, "Error formatting timestamp to date and time string")
+            //Timber.e(e, "Error formatting timestamp to date and time string")
             Pair("", "")
         }
     }
