+++ presentation/src/main/java/com/anytypeio/anytype/presentation/history/VersionHistoryVMFactory.kt
@@ -7,11 +7,9 @@ import com.anytypeio.anytype.domain.history.GetVersions
 import com.anytypeio.anytype.domain.history.SetVersion
 import com.anytypeio.anytype.domain.history.ShowVersion
 import com.anytypeio.anytype.domain.misc.DateProvider
-import com.anytypeio.anytype.domain.misc.LocaleProvider
 import com.anytypeio.anytype.domain.misc.UrlBuilder
 import com.anytypeio.anytype.domain.objects.StoreOfRelations
 import com.anytypeio.anytype.domain.search.SearchObjects
-import com.anytypeio.anytype.presentation.editor.cover.CoverImageHashProvider
 import com.anytypeio.anytype.presentation.editor.render.DefaultBlockViewRenderer
 import com.anytypeio.anytype.presentation.history.VersionHistoryViewModel.VmParams
 import com.anytypeio.anytype.presentation.sets.state.ObjectStateReducer
@@ -22,14 +20,12 @@ class VersionHistoryVMFactory @Inject constructor(
     private val getVersions: GetVersions,
     private val objectSearch: SearchObjects,
     private val dateProvider: DateProvider,
-    private val localeProvider: LocaleProvider,
     private val analytics: Analytics,
     private val urlBuilder: UrlBuilder,
     private val showVersion: ShowVersion,
     private val setVersion: SetVersion,
     private val renderer: DefaultBlockViewRenderer,
     private val setStateReducer: ObjectStateReducer,
-    private val coverImageHashProvider: CoverImageHashProvider,
     private val storeOfRelations: StoreOfRelations
 ) : ViewModelProvider.Factory {
 
@@ -40,14 +36,12 @@ class VersionHistoryVMFactory @Inject constructor(
             getVersions = getVersions,
             objectSearch = objectSearch,
             dateProvider = dateProvider,
-            localeProvider = localeProvider,
             analytics = analytics,
             urlBuilder = urlBuilder,
             showVersion = showVersion,
             renderer = renderer,
             setVersion = setVersion,
             setStateReducer = setStateReducer,
-            coverImageHashProvider = coverImageHashProvider,
             storeOfRelations = storeOfRelations
         ) as T
     }
